Keep track of pending Redis get calls in client side caching to avoid race with invalidations. #3481
